# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "GoogleSheetsExportHistory": {
        "description": "Schema for individual export history entry.",
        "properties": {
          "created_at": {
            "description": "Creation timestamp (ISO 8601)",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "description": "Export record UUID",
            "title": "Id",
            "type": "string"
          },
          "updated_at": {
            "description": "Last update timestamp (ISO 8601)",
            "title": "Last Updated At",
            "type": "string"
          },
          "search_params": {
            "additionalProperties": true,
            "description": "Search parameters snapshot",
            "title": "Search Params",
            "type": "object"
          },
          "spreadsheet_id": {
            "description": "Google Sheets spreadsheet ID",
            "title": "Spreadsheet Id",
            "type": "string"
          },
          "spreadsheet_url": {
            "description": "Shareable URL",
            "title": "Spreadsheet Url",
            "type": "string"
          },
          "total_rows": {
            "description": "Number of rows exported",
            "title": "Total Rows",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "spreadsheet_id",
          "spreadsheet_url",
          "search_params",
          "total_rows",
          "created_at",
          "updated_at"
        ],
        "title": "GoogleSheetsExportHistory",
        "type": "object"
      },
      "GoogleSheetsExportHistoryResponse": {
        "description": "Response schema for export history endpoint.",
        "properties": {
          "exports": {
            "description": "List of user's Google Sheets exports",
            "items": {
              "$ref": "#/components/schemas/GoogleSheetsExportHistory"
            },
            "title": "Exports",
            "type": "array"
          },
          "total": {
            "description": "Total number of exports in history",
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "exports",
          "total"
        ],
        "title": "GoogleSheetsExportHistoryResponse",
        "type": "object"
      },
      "GoogleSheetsExportRequest": {
        "description": "Request schema for Google Sheets export endpoint.\n\nSupports both \"create new spreadsheet\" and \"update existing\" modes.",
        "properties": {
          "licitacoes": {
            "description": "List of procurement contracts to export",
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "maxItems": 10000,
            "minItems": 1,
            "title": "Licitacoes",
            "type": "array"
          },
          "mode": {
            "default": "create",
            "description": "'create' for new spreadsheet, 'update' for existing",
            "pattern": "^(create|update)$",
            "title": "Mode",
            "type": "string"
          },
          "spreadsheet_id": {
            "anyOf": [
              {
                "pattern": "^[a-zA-Z0-9_-]{44}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Google Sheets spreadsheet ID (required for mode='update')",
            "title": "Spreadsheet Id"
          },
          "title": {
            "description": "Spreadsheet title",
            "maxLength": 100,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "licitacoes",
          "title"
        ],
        "title": "GoogleSheetsExportRequest",
        "type": "object"
      },
      "GoogleSheetsExportResponse": {
        "description": "Response schema for successful Google Sheets export.",
        "properties": {
          "spreadsheet_id": {
            "description": "Google Sheets spreadsheet ID",
            "title": "Spreadsheet Id",
            "type": "string"
          },
          "spreadsheet_url": {
            "description": "Full shareable URL to the spreadsheet",
            "title": "Spreadsheet Url",
            "type": "string"
          },
          "success": {
            "default": true,
            "description": "Export success indicator",
            "title": "Success",
            "type": "boolean"
          },
          "total_rows": {
            "description": "Number of contracts exported",
            "minimum": 0.0,
            "title": "Total Rows",
            "type": "integer"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "ISO 8601 timestamp (only for mode='update')",
            "title": "Updated At"
          }
        },
        "required": [
          "spreadsheet_id",
          "spreadsheet_url",
          "total_rows"
        ],
        "title": "GoogleSheetsExportResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "API para busca e an\u00e1lise de licita\u00e7\u00f5es de uniformes no PNCP. Permite filtrar oportunidades por estado, valor e keywords, gerando relat\u00f3rios Excel e resumos executivos via IA.",
    "title": "SmartLic API",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/export/google-sheets": {
      "post": {
        "description": "Export search results to Google Sheets.\n\nCreates a new Google Sheets spreadsheet or updates an existing one\nwith search results. Applies professional formatting (green header,\ncurrency symbols, hyperlinks).\n\nRequest Body:\n    {\n        \"licitacoes\": [...],\n        \"title\": \"SmartLic - Uniformes - 09/02/2026\",\n        \"mode\": \"create\" | \"update\",\n        \"spreadsheet_id\": \"optional for update mode\"\n    }\n\nReturns:\n    {\n        \"success\": true,\n        \"spreadsheet_id\": \"...\",\n        \"spreadsheet_url\": \"https://docs.google.com/spreadsheets/d/...\",\n        \"total_rows\": 142,\n        \"updated_at\": \"2026-02-09T15:30:00Z\"  // only for update mode\n    }\n\nErrors:\n    401 Unauthorized: User hasn't authorized Google Sheets\n    403 Forbidden: Token revoked or insufficient permissions\n    429 Too Many Requests: Google API quota exceeded\n    500 Internal Server Error: Google API error\n\nExample:\n    POST /api/export/google-sheets\n    {\n        \"licitacoes\": [...],\n        \"title\": \"SmartLic - Uniformes - SP, RJ\",\n        \"mode\": \"create\"\n    }",
        "operationId": "export_to_google_sheets_api_export_google_sheets_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoogleSheetsExportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoogleSheetsExportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Export To Google Sheets",
        "tags": [
          "export"
        ]
      }
    },
    "/api/export/google-sheets/history": {
      "get": {
        "description": "Get user's Google Sheets export history.\n\nReturns list of all exports with spreadsheet URLs and search parameters.\nUseful for \"re-open last export\" and usage analytics.\n\nQuery Parameters:\n    limit: Maximum number of exports to return (default: 50, max: 100)\n\nReturns:\n    {\n        \"exports\": [\n            {\n                \"id\": \"uuid\",\n                \"spreadsheet_id\": \"...\",\n                \"spreadsheet_url\": \"https://docs.google.com/...\",\n                \"search_params\": {\"ufs\": [\"SP\"], \"setor\": \"Uniformes\"},\n                \"total_rows\": 142,\n                \"created_at\": \"2026-02-09T15:30:00Z\",\n                \"updated_at\": \"2026-02-09T15:30:00Z\"\n            }\n        ],\n        \"total\": 1\n    }\n\nExample:\n    GET /api/export/google-sheets/history?limit=20",
        "operationId": "get_export_history_api_export_google_sheets_history_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoogleSheetsExportHistoryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Export History",
        "tags": [
          "export"
        ]
      }
    },
...
    "/v1/api/export/google-sheets": {
      "post": {
        "description": "Export search results to Google Sheets.\n\nCreates a new Google Sheets spreadsheet or updates an existing one\nwith search results. Applies professional formatting (green header,\ncurrency symbols, hyperlinks).\n\nRequest Body:\n    {\n        \"licitacoes\": [...],\n        \"title\": \"SmartLic - Uniformes - 09/02/2026\",\n        \"mode\": \"create\" | \"update\",\n        \"spreadsheet_id\": \"optional for update mode\"\n    }\n\nReturns:\n    {\n        \"success\": true,\n        \"spreadsheet_id\": \"...\",\n        \"spreadsheet_url\": \"https://docs.google.com/spreadsheets/d/...\",\n        \"total_rows\": 142,\n        \"updated_at\": \"2026-02-09T15:30:00Z\"  // only for update mode\n    }\n\nErrors:\n    401 Unauthorized: User hasn't authorized Google Sheets\n    403 Forbidden: Token revoked or insufficient permissions\n    429 Too Many Requests: Google API quota exceeded\n    500 Internal Server Error: Google API error\n\nExample:\n    POST /api/export/google-sheets\n    {\n        \"licitacoes\": [...],\n        \"title\": \"SmartLic - Uniformes - SP, RJ\",\n        \"mode\": \"create\"\n    }",
        "operationId": "export_to_google_sheets_v1_api_export_google_sheets_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoogleSheetsExportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoogleSheetsExportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Export To Google Sheets",
        "tags": [
          "export"
        ]
      }
    },
    "/v1/api/export/google-sheets/history": {
      "get": {
        "description": "Get user's Google Sheets export history.\n\nReturns list of all exports with spreadsheet URLs and search parameters.\nUseful for \"re-open last export\" and usage analytics.\n\nQuery Parameters:\n    limit: Maximum number of exports to return (default: 50, max: 100)\n\nReturns:\n    {\n        \"exports\": [\n            {\n                \"id\": \"uuid\",\n                \"spreadsheet_id\": \"...\",\n                \"spreadsheet_url\": \"https://docs.google.com/...\",\n                \"search_params\": {\"ufs\": [\"SP\"], \"setor\": \"Uniformes\"},\n                \"total_rows\": 142,\n                \"created_at\": \"2026-02-09T15:30:00Z\",\n                \"updated_at\": \"2026-02-09T15:30:00Z\"\n            }\n        ],\n        \"total\": 1\n    }\n\nExample:\n    GET /api/export/google-sheets/history?limit=20",
        "operationId": "get_export_history_v1_api_export_google_sheets_history_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoogleSheetsExportHistoryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Export History",
        "tags": [
          "export"
        ]
      }
    },
...
  }
}